Split up AuthenticationService and make `MsaAuthenticationService` extend it
