+++ src/main/java/com/github/steveice10/mc/auth/service/MsaAuthenticationService.java
@@ -1,5 +1,6 @@
 package com.github.steveice10.mc.auth.service;
 
+import com.github.steveice10.mc.auth.exception.request.InvalidCredentialsException;
 import com.github.steveice10.mc.auth.exception.request.RequestException;
 import com.github.steveice10.mc.auth.util.HTTP;
 
@@ -7,43 +8,53 @@ import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 
-public class MsaAuthenticationService extends Service {
-    private String clientId;
-
+public class MsaAuthenticationService extends AuthenticationService {
     private static final URI MS_CODE_ENDPOINT = URI.create("https://login.microsoftonline.com/consumers/oauth2/v2.0/devicecode");
     private static final URI MS_TOKEN_ENDPOINT = URI.create("https://login.microsoftonline.com/consumers/oauth2/v2.0/token");
     private static final URI XBL_AUTH_ENDPOINT = URI.create("https://user.auth.xboxlive.com/user/authenticate");
     private static final URI XSTS_AUTH_ENDPOINT = URI.create("https://xsts.auth.xboxlive.com/xsts/authorize");
     private static final URI MC_LOGIN_ENDPOINT = URI.create("https://api.minecraftservices.com/authentication/login_with_xbox");
 
-    public MsaAuthenticationService(String clientId) {
-        super(URI.create(""));
+    private static final URI EMPTY_URI = URI.create("");
 
-        this.clientId = clientId;
+    private String deviceCode;
+
+    public MsaAuthenticationService(String clientId) {
+        this(clientId, null);
+    }
+
+    public MsaAuthenticationService(String clientId, String deviceCode) {
+        super(clientId, EMPTY_URI);
+
+        this.deviceCode = deviceCode;
     }
 
     public MsCodeResponse getAuthCode() throws RequestException {
-        MsCodeRequest request = new MsCodeRequest(this.clientId);
-        MsCodeResponse response = HTTP.makeRequestForm(this.getProxy(), MS_CODE_ENDPOINT, request.toMap(), MsCodeResponse.class);
-
-        return response;
+        if(this.clientToken == null) {
+            throw new InvalidCredentialsException("Invalid client token.");
+        }
+        MsCodeRequest request = new MsCodeRequest(this.clientToken);
+        return HTTP.makeRequestForm(this.getProxy(), MS_CODE_ENDPOINT, request.toMap(), MsCodeResponse.class);
     }
 
-    public String getSessionTokenFromCode(String deviceCode) throws RequestException {
-        MsTokenRequest request = new MsTokenRequest(this.clientId, deviceCode);
+    public McLoginResponse getLoginResponseFromCode() throws RequestException {
+        if(this.deviceCode == null) {
+            throw new InvalidCredentialsException("Invalid device code.");
+        }
+        MsTokenRequest request = new MsTokenRequest(this.clientToken, this.deviceCode);
         MsTokenResponse response = HTTP.makeRequestForm(this.getProxy(), MS_TOKEN_ENDPOINT, request.toMap(), MsTokenResponse.class);
 
-        return getSessionTokenFromToken(response.access_token);
+        return getLoginResponseFromToken(response.access_token);
     }
 
     /**
-     * Get a Minecraft session token from a
-     * given Microsoft account token
+     * Get a Minecraft login response from the given
+     * Microsoft access token
      *
-     * @param accessToken
-     * @return A session token
+     * @param accessToken the access token
+     * @return The Minecraft login response
      */
-    private String getSessionTokenFromToken(String accessToken) throws RequestException {
+    private McLoginResponse getLoginResponseFromToken(String accessToken) throws RequestException {
         XblAuthRequest xblRequest = new XblAuthRequest(accessToken);
         XblAuthResponse response = HTTP.makeRequest(this.getProxy(), XBL_AUTH_ENDPOINT, xblRequest, XblAuthResponse.class);
 
@@ -51,9 +62,54 @@ public class MsaAuthenticationService extends Service {
         response = HTTP.makeRequest(this.getProxy(), XSTS_AUTH_ENDPOINT, xstsRequest, XblAuthResponse.class);
 
         McLoginRequest mcRequest = new McLoginRequest(response.DisplayClaims.xui[0].uhs, response.Token);
-        McLoginResponse mcResponse = HTTP.makeRequest(this.getProxy(), MC_LOGIN_ENDPOINT, mcRequest, McLoginResponse.class);
+        return HTTP.makeRequest(this.getProxy(), MC_LOGIN_ENDPOINT, mcRequest, McLoginResponse.class);
+    }
 
-        return mcResponse.access_token;
+    @Override
+    public void login() throws RequestException {
+        boolean token = this.clientToken != null && !this.clientToken.isEmpty();
+        boolean device = this.deviceCode != null && !this.deviceCode.isEmpty();
+        boolean password = this.password != null && !this.password.isEmpty();
+        if(!token && !password) {
+            throw new InvalidCredentialsException("Invalid password, device code or access token.");
+        }
+        if(password && (this.username == null || this.username.isEmpty())) {
+            throw new InvalidCredentialsException("Invalid username.");
+        }
+        if(password) {
+            // TODO: Password-based auth to generate token
+        }
+        if(!device) {
+            this.deviceCode = getAuthCode().device_code;
+        }
+        McLoginResponse response = getLoginResponseFromCode();
+        if(response == null) {
+            throw new RequestException("Invalid response received.");
+        }
+        this.username = response.username;
+        this.accessToken = response.access_token;
+        this.loggedIn = true;
+    }
+
+    @Override
+    public void logout() throws RequestException {
+        super.logout();
+        this.clientToken = null;
+    }
+
+    @Override
+    public String toString() {
+        return "MsaAuthenticationService{" +
+                "deviceCode='" + this.deviceCode + '\'' +
+                ", clientToken='" + this.clientToken + '\'' +
+                ", accessToken='" + this.accessToken + '\'' +
+                ", loggedIn=" + this.loggedIn +
+                ", username='" + this.username + '\'' +
+                ", password='" + this.password + '\'' +
+                ", selectedProfile=" + this.selectedProfile +
+                ", properties=" + this.properties +
+                ", profiles=" + this.profiles +
+                '}';
     }
 
     private static class MsCodeRequest {
